main.js разбит на модули (есть warning)
