+++ js/constant.js
@@ -1,0 +1,67 @@
+const MAX_ID_IMG = 25;
+const MAX_ID_COMMMENT = 10000;
+const MAX_COMMMENT = 30;
+const MIN_CNT_LIKE = 15;
+const MAX_CNT_LIKE = 200;
+
+const DESCRIPTION = [
+  'Красивый кадр!',
+  'Отличный ракурс.',
+  'Великолепное фото!',
+  'Очень атмосферно.',
+  'Настоящее искусство.',
+  'Захватывающий момент.',
+  'Стильно и со вкусом.',
+  'Прекрасная цветовая гамма.',
+  'Настроение передано идеально.',
+  'Фотография говорит сама за себя.',
+  'Удивительно красиво.',
+  'Тонко и нежно.',
+  'Создает настроение.',
+  'Энергия в каждом пикселе.',
+  'Так живо и натурально.',
+  'Вдохновляющий снимок!',
+  'Идеальное освещение.',
+  'Очень душевно.',
+  'Отличная композиция.',
+  'Свежий и оригинальный взгляд.',
+  'Люблю такую эстетику.',
+  'Просто супер!',
+  'Такого ещё не видел.',
+  'Мастерство видно сразу.',
+  'Заставляет задуматься.'
+];
+
+const MESSAGE = [
+  'Всё отлично!',
+  'В целом всё неплохо. Но не всё.',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!'
+
+];
+
+const NAME = [
+  'Алексей',
+  'Ирина',
+  'Дмитрий',
+  'Ольга',
+  'Константин',
+  'Мария',
+  'Сергей',
+  'Елена',
+  'Максим',
+  'Наталья',
+  'Владимир',
+  'Татьяна',
+  'Андрей',
+  'Светлана',
+  'Иван',
+  'Егор',
+  'Людмила',
+  'Виктория',
+  'Роман',
+  'Юлия'];
+
+export {MAX_ID_IMG, MAX_ID_COMMMENT, MAX_COMMMENT, MIN_CNT_LIKE, MAX_CNT_LIKE, DESCRIPTION, MESSAGE, NAME};
